pyproject.toml
[project]
name = "telephony-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.4",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.8.0",
    "fastapi>=0.110.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
]
